requirements.txt
python-binance==1.0.10
TA-Lib==0.4.20
numpy==1.20.3
asyncio==3.4.3
websocket-client==1.0.1
asyncio==3.4.3
grequests==0.6.0
